+++ lib/riaction/event_performer.rb
@@ -6,6 +6,7 @@ module Riaction
     @queue = :riaction_event_logger
 
     # Sends an event to IActionable based on the name of a riaction class and the ID used to locate the instance
+    def self.perform(event_name, klass_name, id, attempt=0)
       event_name_sym = event_name.to_sym
       iactionable_api = IActionable::Api.new
       event_object = klass_name.constantize.find_by_id!(id)
@@ -14,11 +15,15 @@ module Riaction
       if (  klass_name.constantize.riactionary? &&
             klass_name.constantize.riaction_events? &&
             klass_name.constantize.riaction_defines_event?(event_name_sym) )
-        iactionable_api.log_event(  event_params[event_name_sym][:profile][:type],
-                                    event_params[event_name_sym][:profile][:id_type],
-                                    event_params[event_name_sym][:profile][:id],
-                                    event_name_sym,
-                                    event_params[event_name_sym][:params])
+        if event_params.has_key?(event_name_sym)
+          iactionable_api.log_event(  event_params[event_name_sym][:profile][:type],
+                                      event_params[event_name_sym][:profile][:id_type],
+                                      event_params[event_name_sym][:profile][:id],
+                                      event_name_sym,
+                                      event_params[event_name_sym][:params])
+        else
+          raise ::Riaction::ConfigurationError.new("Instance of #{klass_name} with ID #{id} could not construct event parameters for event #{event_name}.  Is the profile a valid one?")
+        end
       else
         raise ::Riaction::ConfigurationError.new("#{klass_name} does not define event #{event_name}")
       end